{UNSW Unite}v.0.12.0
Functionality for going/not going implemented for saved events all events and organiser events pages. Organiser events basically done - just need to implement the follow functionality. Past events to be done. Approvals to be done. Following to be done. Dashboards to be done.  Manage Events to be done. View Participants to be done.